Update and renaming of main.py
This version, main[5.0b], is a beta of version 5.0, the difference between  5.0b and 5.0 is that 5.0b just implemented comma support on addition, but on 5.0 it will be supported fully, in addition to that, the comma support algorithm is already tested and working. About version 4.1, let's just forget that it ever existed, because the 5.0 actually is 4.1, but I decided to rename it because I think comma support is a major update.